Backgrounds and Icons
Adding icons for the navigation bar menu. I will be changing the .xml files that are being used to .png due to the fact that .png files are easier to be edited.
Also adding backgrounds for the Moeda screen and for the navigation bar.
Background of Moeda was created considering it can't be polluted and it can't obscure the font.